requirements.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1

# Python PostgreSQL DB adapter


# drf
django-filter==2.2.0
djangorestframework==3.11.0
markdown==3.2.2

# Models
django-model-utils==4.0.0

six==1.15.0

django-oauth-toolkit==1.3.2

django-phonenumber-field==4.0.0
phonenumbers==8.12.4

django-cors-headers==3.3.0



certifi==2020.4.5.1
chardet==3.0.4

idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
oauthlib==3.1.0
requests==2.23.0

urllib3==1.25.9
zipp==3.1.0

coreapi==2.3.3
coreschema==0.0.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
uritemplate==3.0.1

django-debug-toolbar==2.2

# sentry
sentry-sdk==0.15.1

# storage
django-storages==1.9.1
# s3
django-boto==0.3.12

pycodestyle==2.6.0
django-crispy-forms==1.9.1


boto3==1.14.40
